NDEV-2332: Move types from program::evm::tracing::mod to lib::tracing::mod
